+++ pkgs/standards/peagen/peagen/gateway/__init__.py
@@ -27,7 +27,7 @@ from peagen.plugins.queues import QueueBase
 from peagen.transport import RPCDispatcher, RPCRequest
 from peagen.transport.jsonrpc import RPCException
 from peagen.orm import Base, Status, Task
-from peagen.models.schemas import TaskRead
+from peagen.schemas import TaskRead
 from peagen.orm.task.task_run import TaskRun
 
 from peagen.gateway.ws_server import router as ws_router
