+++ pkgs/standards/peagen/peagen/handlers/__init__.py
@@ -5,11 +5,11 @@ from __future__ import annotations
 from typing import Any, Dict
 
 from peagen.orm import Task
-from peagen.models.schemas import TaskRead
+from peagen.schemas import TaskRead
 
 
 def ensure_task(task: Task | Dict[str, Any]) -> TaskRead | Task:
-    """Return ``task`` as a :class:`~peagen.models.schemas.TaskRead` instance."""
+    """Return ``task`` as a :class:`~peagen.schemas.TaskRead` instance."""
 
     if isinstance(task, Task):
         return task
